+++ .github/workflows/deploy-function.yml
@@ -16,34 +16,34 @@ on:
         required: true
         type: choice
         options:
-          - all
-          - process_observation_activity
-          - process_observation_create_analytics
-          - process_observation_update_analytics
-          - process_observation_delete_analytics
-          - process_user
-          - process_invite
-          - process_invite_register
-          - import_wld_data
-          - import_meteoswiss_data
-          - export_meteoswiss_data
-          - create_thumbnails
-          - rollover_phenoyear
-          - process_ts_user
-          - process_ts_observation
-          - process_ts_individual
-          - process_ts_definitions
-          - process_ts_invite
-          - enqueue_individual_map
-          - process_individual_map
-          - process_dragino_http
-          - process_dragino_phaenonet
-          - process_dragino_bq
-          - process_dragino_permarobotics
-          - set_sensor
-          - promote_ranger
-          - test
-          - e2e_clear_individuals
+          # - all
+          # - process_observation_activity
+          # - process_observation_create_analytics
+          # - process_observation_update_analytics
+          # - process_observation_delete_analytics
+          # - process_user
+          # - process_invite
+          # - process_invite_register
+          # - import_wld_data
+          # - import_meteoswiss_data
+          # - export_meteoswiss_data
+          # - create_thumbnails
+          # - rollover_phenoyear
+          # - process_ts_user
+          # - process_ts_observation
+          # - process_ts_individual
+          # - process_ts_definitions
+          # - process_ts_invite
+          # - enqueue_individual_map
+          # - process_individual_map
+          # - process_dragino_http
+          # - process_dragino_phaenonet
+          # - process_dragino_bq
+          # - process_dragino_permarobotics
+          # - set_sensor
+          # - promote_ranger
+          # - test
+          # - e2e_clear_individuals
       action:
         description: Action
         required: true
@@ -99,17 +99,14 @@ jobs:
           - name: process_invite
             entrypoint: process_invite_write
             trigger: --trigger-resource "projects/$PROJECT/databases/(default)/documents/invites/{invite_id}" --trigger-event providers/cloud.firestore/eventTypes/document.write
-          - name: import_wld_data
-            entrypoint: import_wld_data_finalize
-            trigger: --trigger-resource "$PROJECT.appspot.com" --trigger-event google.storage.object.finalize
           - name: import_meteoswiss_data
             entrypoint: ps_import_meteoswiss_data_publish
             trigger: --trigger-resource import_meteoswiss_data --trigger-event google.pubsub.topic.publish
           - name: export_meteoswiss_data
             entrypoint: export_meteoswiss_data_manual
             trigger: --trigger-resource export_meteoswiss_data --trigger-event google.pubsub.topic.publish
-          - name: create_thumbnails
-            entrypoint: create_thumbnail_finalize
+          - name: process_new_files
+            entrypoint: st_appspot_finalize
             trigger: --trigger-resource "$PROJECT.appspot.com" --trigger-event google.storage.object.finalize
           - name: rollover_phenoyear
             entrypoint: rollover_manual
